requirements.txt
Flask==2.2.3
alembic==1.10.3
psycopg2==2.9.6
pytest==7.3.0
pyjwt==2.6.0
Cerberus==1.3.4
Flask_Cors==3.0.10
gunicorn==20.1.0
